requirements.txt
flask
requests
deepl
werkzeug
python-dotenv
